Implement RigidbodyManager and update Sub Prefab.
Complete the RBM class that calculates all of the measurements for
the sub using its Rigidbody, and as a need for testing, created an
Information class to view said measurements.

Also fixed a bug where pressing an input fast enough would cause
the input to seem continuing to be held by updating
SubObjectInputController.CheckInput() as a result of the toggle
method before. Now it sets the bool every update depending on
whether its held or not.

As a result of these changes, I applied them to the sub prefab.