New CameraController
Recreating the design of the camera controller to allow preemptive
changes to positions and rotations of the camera without applying
them until the very end.

Right now it only allows orbiting around the pivot, such that
the pivot never changes its transform, but rather the camera
updates both transformation and rotation.