Add Camera Copy to CameraController2
Created a sort of copy of the camera's transform that allows all
changes to be made to it in methods, then at the very end updates
the transform of the real camera using its field.

As mentioned in a previous commit, this allows a less destructive
way of manipulating the camera's transform without affecting it
directly to prevent unsecure changes.