Remove Reference of Camera Child; Add Comments
CameraController2 no longer requires the reference of the camera
object to be set in the inspector. This prevents user-error and
the script will instead look for the components in the camera
child and contain them in a small private class to be accessed
by methods and later updated using the ghost camera.

Added comments for clarity and partitioning.