Update Camera distance system; Add camera reset with collision.
With the addition of a desired and actual distance, the camera will
try to constantly go towards the desired distance, allowing a
smoothing effect to be performed.

Implemented a method to reset the camera back to its original
position when it's no longer being occluded. Still obeys the rules
of smoothing towards the original position.

There still remains the bug of shearing (esp. with the floor) and
I will investigate its cause and fix it.