Fix Camera Shearing; Update several methods
Shearing was caused by relying on a linecast distance in
ResetToPreOccluded() which, while is the impact point, is
dependent on the shape formed from the pre-occluded distance.
Therefore setting that as the desired distance might cause the
camera to still pierce through geometry as it has not been checked
whether that distance is "safe".

Changed CalculateBetterDistance to CalculateSaferDistance and will
start with an initial point and move towards to the desired point,
checking at each interval if the distance is "safe"

Added a debug method OutputStatus() to help in checking the several
distances of the camera and whether it is in an occluded state.