Add camera switcher; Rename CameraController
Pressing the C button will now switch the camera between a third
and first person perspective, handled by the CameraSwitcher
component added to the CameraPivot game object.

Changed CameraController to CameraThirdPersonController due
to the ambiguity caused when the sim now has two camera to handle.
Because it was in charge of handling the camera, its name was
changed to be more specific about what camera it's supposed to
manage.

In addition to switching between the normal, third person camera
and the first person camera, the script will also disable the
CameraThirdPersonController should it not be needed.