Add camera collision debug scene
Needed to test the third person camera, so made a scene with various
objects to test it. Nothing seemed obviously wrong, except in the
case where the four points of the camera do not collide with any
object from the origin, but if there is an object between the points
on the rectangle, it might sheer.

A fix to that would be adding four more linecasts of the four
corners, but that adds more overhead and is unnecessary right now
as it doesn't add much value.

Also created a prefab for the Container game object.