Add realistic physics to sub
Added Motor, SoloMotor, MotorVisuals, and PhysicsSim classes.

The Motor class contains info about each motor, the SoloMotor class
contains allows a motor to be added to a game object, and the PhysicsSim
class calculates movement based on all the SoloMotors selected.

The MotorVisuals class as well as the added assets control the look of
the motors.