Update Submarine prefab
Added CameraManager (which exists in Sub prefab) to Submarine
prefab so that third and person camera controls are possible.

Before, as seen in PhysicsScene, a CameraDolly was created to
have temporary camera controls, but this was only for that branch
and will not be included in MainScene as the CameraManager is the
actual system to be used.