Update distortion flow
Fixed on bug that the shader would ignore transparent materials
since they would be rendered after the shader had been applied,
hence not affecting such materials.

However, this now introduces the new issue where transparent
materials aren't rendered at all when the shader affects them,
making them 100% see through the water surface (where the shader
takes effect).

--------

Added a new modifier to the uvOffset (which affects how refracted
the objects underwater become) that takes into account the distance
the camera is from the water surface.

The modifier is based on a logistic function that, when graphed,
will start at ~0 when the camera distance from surface is 0,
increase to 0.5 when the camera is 5 units away, and cap out at
~1 as camera distance increases.

This prevents objects from being refracted too harshly when the
camera is really close to the water surface. Instead, the objects
will appear relatively normal.