Create Toggler script
Util script that offers a method to toggle the active state of
the game object it's attached to. Created because
MonoBehvaior.SetActive() can only set with some value, but not
be able to toggle it. Buttons such as the SettingsMenu toggles
the UI element, so this script is exactly for that.