Remove BoxCollider from Container
The camera's collision system messes up when the sub is at the
water line. This was a result of the
CameraThirdPersonController.GetOcclusion() using Physics.Linecast()
which depends on game objects having a collider, and the
WaterVolume object in the scene has a Box Collider.

The Container prefab was created in 5c1eedf when the camera
collision was being further tested, so possibly the box collider
may have been added by mistake.

This commit removes the box collider and updates the prefab to
reflect this. Now the camera won't break near the water line.